fix(frontend): keep decrypted preview when effect re-fires without password
A successful decrypt calls setIsDecrypted(true), which changes a fetchFile
dependency and re-runs the effect with no password. The old guard fell through
and loadPlaintextBytes threw for the missing password, replacing the working
preview with a bogus "invalid password" error. Return early when already
decrypted, and extract the decode helpers into decode.ts.
